style: Node class를 별도의 파일로 분리
